+++ packages/terra-slide-panel/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed bug where `terra-slide-panel` crashes when open by default on load.
+
 ## 3.38.0 - (March 29, 2023)
 
 * Changed